Moved engine exposer to dedicated class
