+++ engine/Controller/Lua/LuaGameBridge.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "LuaManager.h"
+
+class LuaGameBridge
+{
+public:
+	static void ExposeEngine(LuaManager* luaState);
+};
